refactor: move global constants to cloud pkg
