--- internal/agent/globals/constants.go
@@ -1,16 +1,0 @@
-package globals
-
-const (
-	ConfigUpdatedAtMethod               = "GET"
-	ConfigUpdatedAtAPI                  = "/config"
-	ConfigAPIMethod                     = "GET"
-	ConfigAPI                           = "/api/runtime/config"
-	ListsAPIMethod                      = "GET"
-	ListsAPI                            = "api/runtime/firewall/lists"
-	EventsAPIMethod                     = "POST"
-	EventsAPI                           = "/api/runtime/events"
-	MinHeartbeatIntervalInMS            = 120000
-	MaxAttackDetectedEventsPerInterval  = 100
-	AttackDetectedEventsIntervalInMs    = 60 * 60 * 1000 // 1 hour
-	MinStatsCollectedForRelevantMetrics = 10000
-)
